Feature/working memory dry 2 (#140)
* Moved   dry_run_all_pipes() out of PipeLibrary, it's a static func called (or not) by make()

* Added dry_run_pipes in its noew module

* Nicer log.info with indentation when running pipes

* Improve Pipe log info indent
