+++ pipelex/pipe_controllers/pipe_controller.py
@@ -3,6 +3,7 @@ from typing import Optional
 
 from typing_extensions import override
 
+from pipelex import log
 from pipelex.core.pipe_abstract import PipeAbstract, PipeType
 from pipelex.core.pipe_output import PipeOutput
 from pipelex.core.pipe_run_params import PipeRunMode, PipeRunParams
@@ -31,6 +32,11 @@ class PipeController(PipeAbstract):
 
         match pipe_run_params.run_mode:
             case PipeRunMode.LIVE:
+                indent_level = len(pipe_run_params.pipe_stack) - 1
+                indent = "   " * indent_level
+                label = f"{indent}{self.class_name}: {self.code}".ljust(60)
+                output = self.output_concept_code.split(".")[-1]
+                log.info(f"{label} → {output}")
                 pipe_output = await self._run_controller_pipe(
                     job_metadata=job_metadata,
                     working_memory=working_memory,
@@ -38,6 +44,12 @@ class PipeController(PipeAbstract):
                     output_name=output_name,
                 )
             case PipeRunMode.DRY:
+                name = f"Dry {self.class_name}"
+                indent_level = len(pipe_run_params.pipe_stack) - 1
+                indent = "   " * indent_level
+                label = f"{indent}{name}: {self.code}".ljust(60)
+                output = self.output_concept_code.split(".")[-1]
+                log.info(f"{label} → {output}")
                 pipe_output = await self._dry_run_controller_pipe(
                     job_metadata=job_metadata,
                     working_memory=working_memory,
